Sejin

+++ 2025_SMU_3D_Game_Task/Assets/Scripts/Player/PlayerHealth.cs
@@ -50,6 +50,6 @@ public class PlayerHealth : MonoBehaviour
     {
         curHealthTxt.text = curHealth.ToString();
         maxHealthTxt.text = maxHealth.ToString();
-        healthSlider.value = curHealth / maxHealth;
+        healthSlider.value = (float)curHealth / (float)maxHealth;
     }
 }
